+++ example/web_frontend/src/lib/assets/js/model/sample.js
@@ -1,0 +1,12 @@
+export default (() => {
+    return new Promise(async (resolve, reject) => {
+        try {
+            let lib = {};
+            console.log("sample.js say hello!");
+
+            resolve(lib);
+        } catch (error) {
+            reject(error);
+        }
+    });
+})();