+++ Changelog.md
@@ -13,11 +13,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Create sample package.json.example which allow user self install node modules to components. Completed on 2024-04-22
+
 ### Changed
+
+- Move code from `src/index.js` to `index.js` to reduce a js file. Completed on 2024-04-22
+- Move `src/coresetting.toml.example` to parent folder. Completed on 2024-04-22
+- Change `src/coresetting.toml.example` to parent folder. Completed on 2024-04-22
+- Change the design in `init,prepare_rules` functions in `index.js`. Completed on 2024-04-22
 
 ### Deprecated
 
 ### Removed
+
+- Delete `src/index.js`. Completed on 2024-04-23
 
 ### Fixed